+++ src/modules/matrix.js
@@ -1,4 +1,6 @@
-import { WrongMatrixStructureError } from '../errors.js';
+import { checktypes } from 'es7-checktypes-decorator';
+
+import { WrongSchemaError } from '../errors.js';
 import { arrayEqual, iter } from '../reusables.js';
 
 /**
@@ -14,23 +16,30 @@ class Matrix {
         this.name = 'matrix';
     }
 
+    @checktypes('DataFrame')
     /**
      * Check if two DataFrames are commutative, if both have the same dimensions.
-     * @param {Array} dfDim The second DataFrame dim to check.
+     * @param {DataFrame} df The second DataFrame to check.
+     * @param {Boolean} [reverse = false] Revert the second DataFrame before the comparison.
      * @returns {Boolean} True if they are commutative, else false.
+     * @example
+     * df.matrix.isCommutative(df2)
      */
-    isCommutative(dfDim) {
-        return arrayEqual(this.df.dim(), dfDim, true);
+    isCommutative(df, reverse = false) {
+        return arrayEqual(this.df.dim(), reverse ? df.dim().reverse() : df.dim(), true);
     }
 
+    @checktypes('DataFrame')
     /**
      * Provide an elements pairwise addition of two DataFrames having the same dimensions.
      * @param {DataFrame} df The second DataFrame to add.
      * @returns {DataFrame} A new DataFrame resulting to the addition two DataFrames.
+     * @example
+     * df.matrix.add(df2)
      */
     add(df) {
-        if (!this.isCommutative(df.dim())) {
-            throw new WrongMatrixStructureError(this.df.dim(), df.dim());
+        if (!this.isCommutative(df)) {
+            throw new WrongSchemaError(this.df.dim(), df.dim());
         }
         const columns = [...Array(this.df.dim()[1]).keys()];
         return this.df.__newInstance__([...iter(
@@ -43,25 +52,30 @@ class Matrix {
         )], this.df.listColumns());
     }
 
+    @checktypes('Number')
     /**
      * Provide a scalar product between a number and a DataFrame.
      * @param {Number} number The number to multiply.
      * @returns {DataFrame} A new DataFrame resulting to the scalar product.
+     * @example
+     * df.matrix.product(6)
      */
     product(number) {
         return this.df.map(row => row.toArray().map(column => column * number));
     }
 
+    @checktypes('DataFrame')
     /**
      * Multiply one DataFrame n x p and a second p x n.
      * @param {DataFrame} df The second DataFrame to multiply.
      * @returns {DataFrame} A new n x n DataFrame resulting to the product of two DataFrame.
+     * @example
+     * df.matrix.dot(df)
      */
     dot(df) {
-        if (!this.isCommutative(df.dim().reverse())) {
-            throw new WrongMatrixStructureError(this.df.dim(), df.dim().reverse());
+        if (!this.isCommutative(df, true)) {
+            throw new WrongSchemaError(this.df.dim(), df.dim().reverse());
         }
-        const transposedDF = df.transpose();
         const columns = [...Array(this.df.dim()[0]).keys()];
         return this.df.__newInstance__([...iter(
             Object.keys([...this.df]),
@@ -70,7 +84,7 @@ class Matrix {
                 return [...iter(
                     columns,
                     column => {
-                        const b = [...transposedDF][column].toArray();
+                        const b = [...df.transpose()][column].toArray();
                         return Object.keys(b).reduce((p, n) => p + b[n] * a[n], 0);
                     }
                 )];